remove tangle between extHost.protocol and extHostTunnelService, fyi @roblourens
