+++ src/vs/workbench/api/common/extHostTunnelService.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { ExtHostTunnelServiceShape } from 'vs/workbench/api/common/extHost.protocol';
+import { ExtHostTunnelServiceShape, TunnelDto } from 'vs/workbench/api/common/extHost.protocol';
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
 import * as vscode from 'vscode';
 import { ProvidedPortAttributes, RemoteTunnel, TunnelCreationOptions, TunnelOptions, TunnelPrivacyId } from 'vs/platform/tunnel/common/tunnel';
@@ -13,15 +13,7 @@ import { IExtHostRpcService } from 'vs/workbench/api/common/extHostRpcService';
 import { IExtensionDescription } from 'vs/platform/extensions/common/extensions';
 import { CandidatePort } from 'vs/workbench/services/remote/common/remoteExplorerService';
 
-export interface TunnelDto {
-	remoteAddress: { port: number; host: string };
-	localAddress: { port: number; host: string } | string;
-	public: boolean;
-	privacy: TunnelPrivacyId | string;
-	protocol: string | undefined;
-}
-
-export namespace TunnelDto {
+export namespace TunnelDtoConverter {
 	export function fromApiTunnel(tunnel: vscode.Tunnel): TunnelDto {
 		return {
 			remoteAddress: tunnel.remoteAddress,
